requirements.txt
Django==3.0.4
mysqlclient==1.4.6
djangorestframework==3.11.0
django-bootstrap-form==3.4
Pillow